repromodel_core/requirements.txt
absl-py==2.1.0
albumentations==1.4.6
annotated-types==0.6.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
coverage==7.5.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
easydict==1.13
easypip==1.3.0
efficientnet_pytorch==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
fsspec==2024.3.1
gdown==5.2.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
ollama==0.2.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pretrainedmodels==0.7.4
pretty-errors==1.2.25
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==0.9.2
torch==2.3.0
torch-tb-profiler==0.4.3
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.18.1
